requirements.txt
djangorestframework-simplejwt == 4.6.0
djangorestframework == 3.12.2
Django == 3.0